Feature/hyeonseung/item front item#9

See merge request cloud_track/class_04/web_project2/team03/spring-boots!212

+++ src/main/java/com/spring_boots/spring_boots/item/entity/Item.java
@@ -33,7 +33,7 @@ public class Item extends BaseTimeEntity {
     private String itemName;
 
     @Column(name = "item_price")
-    private Integer itemPrice;
+    private Long itemPrice;
 
     @Column(name = "item_description")
     private String itemDescription;
